[Manifold Garden] Update level pointer path
This pointer path is based on the LevelLoader::currentLevelSystem pointer to a LevelSystems instance with a m_levelIndex member variable. It should hopefully be more stable than the other path, because I technically do not know what that was pointing at, and on rare occasion it seemed to not work.